Add arp address source
Add the capability to lookup the instance address from the local arp
table on linux.

Add the 'address_source' setting to specify how to lookup instance
addresses. Supported methods are 'agent', 'leases', and 'arp'.  The
libvirt address-interfaces is queried for the 'agent' and 'leases'
sources, but we read the arp cache (in /proc/net/arp) for the 'arp'
method since some systems do not support the interface-addresses 'arp'
source.

The 'address_source' setting defaults to 'agent'. The 'address_source'
can be specified in the [site] settings for a system-wide setting, and
can also be specified in the template definitions, to specify the source
on a per os basis.