Add more santity checks before building and cloning
Before building/cloning, ensure we to not already have a domain
in the way that has no metadata. This can happen if someone makes
a domain with out virt-up that just happens to have a name conflict
or if the metadata file was removed or virt-up failed to create
the metadata file.

Also do not just remove an existing orphaned image file, it could
be for some other image, and not because virt-install failed
previously.

Change Instance.exists() to return true only if both the domain
exists and the metadata file exists. If the metadata file does
not exist, assume that domain is not a virt-up managed domain.