Set the user password with useradd
Running chpasswd to set the user's password is problematic since
it might not be available, and on some systems needs to be run
at first boot.

Instead, set the password when the user account is created with
useradd.  useradd requires a hashed password, so use the crypt
module to hash it with the SHA512 method.