Generate an ansible inventory file
Generate an ansible inventory file with connection host variables for
the cloned instances.  This file is regenerated each time an instance is
cloned or deleted.  The template instances (built with build() are not
included.

Hosts are placed in the 'all' and 'virt_up_managed' groups.  The
inventory file is saved as `$XDG_DATA_HOME/virt-up/inventory.yaml`,
where $XDG_DATA_HOME defaults to `$HOME/.local/shared/`.

Add the --no-inventory option to the virt-up command to exclude
instances from the generated inventory file.

Example:

    $ virt-up test01 -t generic-centos-7
    $ virt-up test02 -t generic-centos-8

    $ ANSIBLE_INVENTORY=myhosts.ini,~/.local/share/virt-up/inventory.yaml
    $ ansible-playbook myplay.yaml