Wait for guest to settle when obtaining IP address
When obtaining the IP address, try to ping it and retry for a period
of time.  This allows both the guest system and the 'address-source' to
settle in case the 'address-source' has a stale entry.

This problem was discovered with DHCP.  The DHCP service in an ASUS
router holds onto the leases and can have multiple IP addresses leased
to the same MAC address.  When obtaining the IP address, the DNS lookup
would return a stale address before the guest system had actually
obtained an address for the current instance.