Wait for guest to settle when obtaining IP address
When obtaining the IP address with arp or dns, try to ping it before
accepting the address with the retry number of attempts.

This problem was discovered with DHCP.  The DHCP service in an router
holds onto the leases and can have multiple IP addresses leased to the
same MAC address.  When obtaining the IP address, the DNS lookup would
return a stale address before the guest system had actually obtained an
address for the current instance.