Add config file test fixture
Add a pytest fixture to create a temp dir with virt-up configuration
files, including the firstboot scripts and playbooks, in preparation of
avoiding reading this data from config.py strings in a future commit.

The lines marked 'deprecated' in config.py will be removed after we've
completed the conversion from read_string().