Add virt-up --init option
Add the virt-up --init option to create default configuration files.

Creates the cfg files (ini format), firstboot shell scripts, and sample
playbooks in the users VIRTUP_CONFIG_HOME directory. If running as root,
creates the files in the global /etc/virt-up directory instead.

This commit completes the move of the playbooks from the module
directory, to the config.py along with the other default config files.