Add configuration support for memory, vcpus, and graphics
These options could be set on the command line, but the defaults were
hardcoded. With this change, memory, vcpus, and graphics default values
are read from the config files when creating a new template image. First
we check the template specific section, and if not found, check the
"common" section of the settings.cfg file.

Regular instances still inherit the default values from the parent
template meta-data.