Default user to current username instead of virt
If the --user command line option is not specified, and the 'user'
config option is not specified, default to the value of the current user
running virt-up instead of 'virt'.

To be consistent with the command line option name, check for 'user' in
the config files. Fallback to 'username' for compat.