Support non-filename template names
Change the Instance.build() arguments in order to support template names
which are not filenames (e.g.  'generic/centos8').

Originally, build() was envisioned to create regular instances, with a
required target and template name. However, this was confusing and error
prone.  Change the build() args to require a template name and unless a
target name is given, derive the instance name from the template name.
The derived name will be safe to use for filenames, even if the template
name contains characters like '/'.

Unfortunately, this is not a backward compatible change.