Remove virt-up init subcommand
The virt-up init subcommand was intended to provide users with a
starting point the first time virt-up is installed, but users would run
it after upgrades and try to manually merge changes, which is confusing
and error prone.

Instead, just remove the virt-up init misfeature.  Users or packagers
must provide working configuration for site-specific setups and
virt-builder images they care about.  Later, we can provide some
examples in the docs.

This also means we do not need to update the virt-up package as new
versions of virt-builder images are released.

Move the example configuration templates to the unit tests, so the unit
tests continue to work. These templates are not exported in the virt-up
packages and are only used for unit testing. Update the packaging to
remove 'package' files and the cookiecutter dependency.